Move google distribution files to google package
